Move translations utils from amo.utils to mkt.translations.utils (bug 1075673)
