+++ apps/amo/tests/test_amo_utils.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import collections
 import os
 import tempfile
 import unittest
@@ -8,17 +7,12 @@ import unittest
 from django.conf import settings
 from django.core.cache import cache
 from django.core.validators import ValidationError
-from django.utils import translation
 
 import mock
-from nose.tools import assert_raises, eq_, ok_, raises
+from nose.tools import assert_raises, eq_, raises
 
-import amo
-from amo.utils import (attach_trans_dict, cache_ns_key, escape_all,
-                       find_language, LocalFileStorage, no_translation,
-                       resize_image, rm_local_tmp_dir, slugify, slug_validator,
-                       to_language)
-from mkt.webapps.models import Webapp
+from amo.utils import (cache_ns_key, escape_all, LocalFileStorage,
+                       resize_image, rm_local_tmp_dir, slugify, slug_validator)
 
 
 u = u'Ελληνικά'
@@ -75,50 +69,6 @@ def test_resize_transparency():
     finally:
         if os.path.exists(dest):
             os.remove(dest)
-
-
-def test_to_language():
-    tests = (('en-us', 'en-US'),
-             ('en_US', 'en-US'),
-             ('en_us', 'en-US'),
-             ('FR', 'fr'),
-             ('el', 'el'))
-
-    def check(a, b):
-        eq_(to_language(a), b)
-    for a, b in tests:
-        yield check, a, b
-
-
-def test_find_language():
-    tests = (('en-us', 'en-US'),
-             ('en_US', 'en-US'),
-             ('en', 'en-US'),
-             ('cy', 'cy'),  # A hidden language.
-             ('FR', 'fr'),
-             ('es-ES', None),  # We don't go from specific to generic.
-             ('xxx', None))
-
-    def check(a, b):
-        eq_(find_language(a), b)
-    for a, b in tests:
-        yield check, a, b
-
-
-def test_no_translation():
-    """
-    `no_translation` provides a context where only the default
-    language is active.
-    """
-    lang = translation.get_language()
-    translation.activate('pt-br')
-    with no_translation():
-        eq_(translation.get_language(), settings.LANGUAGE_CODE)
-    eq_(translation.get_language(), 'pt-br')
-    with no_translation('es'):
-        eq_(translation.get_language(), 'es')
-    eq_(translation.get_language(), 'pt-br')
-    translation.activate(lang)
 
 
 class TestLocalFileStorage(unittest.TestCase):
@@ -278,73 +228,3 @@ class TestEscapeAll(unittest.TestCase):
         eq_(res['dict'], {'x': expected})
         eq_(res['list'], [expected])
         eq_(res['bool'], True)
-
-
-class TestAttachTransDict(amo.tests.TestCase):
-    """
-    Tests for attach_trans_dict. For convenience, we re-use Webapp model
-    instead of mocking one from scratch and we rely on internal Translation
-    unicode implementation, because mocking django models and fields is just
-    painful.
-    """
-
-    def test_basic(self):
-        addon = amo.tests.app_factory(
-            name='Name', description='Description <script>alert(42)</script>!',
-            homepage='http://home.pa.ge', privacy_policy='Policy',
-            support_email='sup@example.com', support_url='http://su.pport.url')
-        addon.save()
-
-        # Quick sanity checks: is description properly escaped? The underlying
-        # implementation should leave localized_string un-escaped but never use
-        # it for __unicode__. We depend on this behaviour later in the test.
-        ok_('<script>' in addon.description.localized_string)
-        ok_('<script>' not in addon.description.localized_string_clean)
-        ok_('<script>' not in unicode(addon.description))
-
-        # Attach trans dict.
-        attach_trans_dict(Webapp, [addon])
-        ok_(isinstance(addon.translations, collections.defaultdict))
-        translations = dict(addon.translations)
-
-        # addon.translations is a defaultdict.
-        eq_(addon.translations['whatever'], [])
-
-        # No-translated fields should be absent.
-        ok_(None not in translations)
-
-        # Build expected translations dict.
-        expected_translations = {
-            addon.privacy_policy_id: [
-                ('en-us', unicode(addon.privacy_policy))],
-            addon.description_id: [
-                ('en-us', unicode(addon.description))],
-            addon.homepage_id: [('en-us', unicode(addon.homepage))],
-            addon.name_id: [('en-us', unicode(addon.name))],
-            addon.support_email_id: [('en-us', unicode(addon.support_email))],
-            addon.support_url_id: [('en-us', unicode(addon.support_url))]
-        }
-        eq_(translations, expected_translations)
-
-    def test_multiple_objects_with_multiple_translations(self):
-        addon = amo.tests.app_factory()
-        addon.description = {
-            'fr': 'French Description',
-            'en-us': 'English Description'
-        }
-        addon.save()
-        addon2 = amo.tests.app_factory(description='English 2 Description')
-        addon2.name = {
-            'fr': 'French 2 Name',
-            'en-us': 'English 2 Name',
-            'es': 'Spanish 2 Name'
-        }
-        addon2.save()
-        attach_trans_dict(Webapp, [addon, addon2])
-        eq_(set(addon.translations[addon.description_id]),
-            set([('en-us', 'English Description'),
-                 ('fr', 'French Description')]))
-        eq_(set(addon2.translations[addon2.name_id]),
-            set([('en-us', 'English 2 Name'),
-                 ('es', 'Spanish 2 Name'),
-                 ('fr', 'French 2 Name')]))
