+++ mkt/feed/indexers.py
@@ -2,12 +2,11 @@
 Indexers for FeedApp, FeedBrand, FeedCollection, FeedShelf, FeedItem for
 feed homepage and curation tool search.
 """
-from amo.utils import attach_trans_dict
-
 import mkt.carriers
 import mkt.feed.constants as feed
 import mkt.regions
 from mkt.search.indexers import BaseIndexer
+from mkt.translations.models import attach_trans_dict
 from mkt.translations.utils import format_translation_es
 from mkt.webapps.models import Webapp
