+++ mkt/purchase/models.py
@@ -8,15 +8,14 @@ import tower
 from babel import Locale, numbers
 from jingo import env
 from jinja2.filters import do_dictsort
-from tower import ugettext as _
 
 import amo
 from amo.helpers import urlparams
-from amo.utils import get_locale_from_lang
 from mkt.site.fields import DecimalCharField
 from mkt.site.helpers import absolutify
-from mkt.site.mail import send_mail, send_mail_jinja
+from mkt.site.mail import send_mail
 from mkt.site.models import ModelBase
+from mkt.translations.utils import get_locale_from_lang
 
 
 class ContributionError(Exception):
