+++ mkt/purchase/tasks.py
@@ -4,10 +4,11 @@ import jingo.helpers
 from celeryutils import task
 from tower import ugettext as _
 
-from amo.utils import get_locale_from_lang
 from mkt.purchase.models import Contribution
 from mkt.site.helpers import absolutify
 from mkt.site.mail import send_html_mail_jinja
+from mkt.translations.utils import get_locale_from_lang
+
 
 log = logging.getLogger('z.purchase.webpay')
 notify_kw = dict(default_retry_delay=15,  # seconds
